frontend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
supabase==2.18.1
python-dotenv==1.1.1
waitress==3.0.2
requests
gradio_client